refactoring test
