+++ index.js
@@ -1,2 +1,2 @@
 module.exports = require('./lib/connect-oembed');
-module.exports.express = require('./lib/express-oembed');
+
